resources/requirements/requirements-dev.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
astroid==3.3.9
attrs==25.3.0
aws-sam-translator==1.95.0
aws-xray-sdk==2.14.0
boto3==1.37.13
botocore==1.37.13
certifi==2025.1.31
cffi==1.17.1
cfn-lint==1.30.0
charset-normalizer==3.4.1
cryptography==44.0.2
dill==0.3.9
docker==7.1.0
graphql-core==3.2.6
idna==3.10
iniconfig==2.0.0
isort==6.0.1
Jinja2==3.1.6
jmespath==1.0.1
joserfc==1.0.4
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
lazy-object-proxy==1.10.0
MarkupSafe==3.0.2
mccabe==0.7.0
moto==5.1.1
mpmath==1.3.0
multipart==1.2.1
networkx==3.4.2
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
packaging==24.2
pathable==0.4.4
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
py-partiql-parser==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pylint==3.3.5
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
responses==0.25.7
rfc3339-validator==0.1.4
rpds-py==0.23.1
s3transfer==0.11.4
setuptools==76.0.0
six==1.17.0
sympy==1.13.3
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
xmltodict==0.14.2
